Include the sensor definition.
The sensor definition doesn't tend to change,
however, it's useful to have for sql queries. Rather
than deal with joins against another table, just put
the sensor definition in with the sensor data.

Deal with an sql join later when there is a second table
with all of the static information (such as human readable
description, sensor definition, sensor type, etc).